pyproject.toml
[tool.poetry]
name = "multable"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
pydantic-argparse = "^0.5.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-isort = "^4.1.1"
black = "^22.3.0"
flakeheaven = "^0.11.1"
flake8-bugbear = "^22.4.25"
pep8-naming = "^0.12.1"
flake8-docstrings = "^1.6.0"
flake8-comprehensions = "^3.10.0"
flake8-eradicate = "^1.2.1"
flake8-mutable = "^1.2.0"
flake8-builtins = "^1.5.3"
flake8-pytest-style = "^1.6.0"
flake8-quotes = "^3.3.1"
darglint = "^1.8.1"
pylint = "^2.13.9"
mypy = "^0.950"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
skip-string-normalization = true

[tool.isort]
profile = "black"

[tool.flakeheaven]
exclude = ["example.py"]
format = "colored"
max_line_length = 88
show_source = false
max_complexity = 10

[tool.flakeheaven.plugins]
pyflakes = ["+*"]
pycodestyle = ["-E203"]
mccabe = ["+*"]
flake8-isort = ["+*"]
flake8-bugbear = ["+*"]
pep8-naming = ["+*"]
flake8-docstrings = ["-*"]
flake8-darglint = ["-*"]
flake8-builtins = ["+*"]
flake8-comprehensions = ["+*"]
flake8-eradicate = ["+*"]
flake8-pytest-style = ["+*"]
flake8-quotes = ["+*"]
flake8-mutable = ["+*"]
pylint = ["-*"]
